Revert "Auto form"
