+++ package.json
@@ -5,15 +5,13 @@
   "dependencies": {
     "@material-ui/core": "^1.1.0",
     "express": "^4.16.3",
-    "next": "^6.0.3",
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
     "react-scripts": "1.1.4"
   },
   "scripts": {
-    "dev": "node server.js",
-    "build": "next build",
-    "start": "NODE_ENV=production node server.js",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   }
