--- pages/signup.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { signupForm } from '../config/forms';
-import AccountCard from '../src/components/AccountCard';
-import withRoot from '../src/withRoot';
-
-export default withRoot(() => <AccountCard form={signupForm} />);
