--- src/getPageContext.js
@@ -1,47 +1,0 @@
-import { SheetsRegistry } from 'jss';
-import {
-  createMuiTheme,
-  createGenerateClassName
-} from '@material-ui/core/styles';
-import purple from '@material-ui/core/colors/purple';
-import lightBlue from '@material-ui/core/colors/lightBlue';
-
-// A theme with custom primary and secondary color.
-// It's optional.
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: lightBlue[100],
-      main: lightBlue[200],
-      dark: lightBlue[300],
-      contrastText: 'white'
-    }
-  }
-});
-
-function createPageContext() {
-  return {
-    theme,
-    // This is needed in order to deduplicate the injection of CSS in the page.
-    sheetsManager: new Map(),
-    // This is needed in order to inject the critical CSS.
-    sheetsRegistry: new SheetsRegistry(),
-    // The standard class name generator.
-    generateClassName: createGenerateClassName()
-  };
-}
-
-export default function getPageContext() {
-  // Make sure to create a new context for every server-side request so that data
-  // isn't shared between connections (which would be bad).
-  if (!process.browser) {
-    return createPageContext();
-  }
-
-  // Reuse context on the client-side.
-  if (!global.__INIT_MATERIAL_UI__) {
-    global.__INIT_MATERIAL_UI__ = createPageContext();
-  }
-
-  return global.__INIT_MATERIAL_UI__;
-}
