--- src/withRoot.js
@@ -1,68 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { MuiThemeProvider } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import getPageContext from './getPageContext';
-import { Grid, colors } from '@material-ui/core';
-
-function withRoot(Component) {
-  class WithRoot extends React.Component {
-    constructor(props) {
-      super(props);
-
-      this.pageContext = this.props.pageContext || getPageContext();
-    }
-
-    componentDidMount() {
-      // Remove the server-side injected CSS.
-      const jssStyles = document.querySelector('#jss-server-side');
-      if (jssStyles && jssStyles.parentNode) {
-        jssStyles.parentNode.removeChild(jssStyles);
-      }
-    }
-
-    pageContext = null;
-
-    render() {
-      // MuiThemeProvider makes the theme available down the React tree thanks to React context.
-      return (
-        <MuiThemeProvider
-          theme={this.pageContext.theme}
-          sheetsManager={this.pageContext.sheetsManager}
-        >
-          {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-          <CssBaseline />
-          <Grid
-            style={{
-              background: `radial-gradient(${colors.lightBlue[50]}, ${
-                colors.lightBlue[300]
-              })`,
-              minHeight: '100vh'
-            }}
-            container
-            justify="center"
-            alignItems="center"
-          >
-            <Component {...this.props} />
-          </Grid>
-        </MuiThemeProvider>
-      );
-    }
-  }
-
-  WithRoot.propTypes = {
-    pageContext: PropTypes.object
-  };
-
-  WithRoot.getInitialProps = ctx => {
-    if (Component.getInitialProps) {
-      return Component.getInitialProps(ctx);
-    }
-
-    return {};
-  };
-
-  return WithRoot;
-}
-
-export default withRoot;
